requirements.txt
python3-wget
dataclasses
lxml
git+https://github.com/daemon/pywikiclean
requests
tqdm
mysql-connector